requirements.txt
torch
pandas
scikit-learn
pyyaml
streamlit
matplotlib
joblib
numpy